Resolve conflict in `ErrorCodes.cpp`: master added codes 766-770, so
`NO_QUERY_TO_KILL` is renumbered to 771.

+++ src/Core/MySQL/PacketsProtocolText.cpp
@@ -244,7 +244,7 @@ ColumnDefinition getColumnDefinition(const String & column_name, const DataTypeP
             charset = CharacterSet::utf8_general_ci;
             break;
     }
-    return ColumnDefinition(column_name, charset, 0, column_type, flags, decimals);
+    return ColumnDefinition(column_name, charset, 0, column_type, static_cast<uint16_t>(flags), decimals);
 }
 
 }
