+++ src/Functions/FunctionDynamicAdaptor.cpp
@@ -232,11 +232,11 @@ ColumnPtr ExecutableFunctionDynamicAdaptor::executeImpl(const ColumnsWithTypeAnd
             {
                 if (byte)
                 {
-                    nested_filter.push_back(0);
+                    nested_filter.push_back(static_cast<UInt8>(0));
                 }
                 else
                 {
-                    nested_filter.push_back(1);
+                    nested_filter.push_back(static_cast<UInt8>(1));
                     ++size_hint;
                 }
             }
